Fixed Updating Of Text
