+++ src/components/canvas/GetSelectedText.jsx
@@ -1,16 +1,122 @@
+import React, { useEffect, useState } from 'react';
 import { track, useEditor } from 'tldraw';
-import React from 'react';
+import { generateQuestions, generateAnswers, generateIdeas } from './GenerateLLM';
+import { createShapeId } from '@tldraw/tldraw';
 
-// Component to get selected texts
+// Function to handle generating items and positioning them
+const handleGenerateItems = async (editor, selectedText, generateFunction, numItems) => {
+  console.log('Generating items for selected text:', selectedText);
+  if (!selectedText) {
+    console.error('No text found for generating items.');
+    return;
+  }
+
+  try {
+    const items = await generateFunction(selectedText, numItems);
+    console.log('Generated items:', items);
+
+    if (items.length > 0) {
+      const selectedShapes = editor.getSelectedShapes();
+      console.log('Selected shapes:', selectedShapes);
+      if (selectedShapes.length === 0) {
+        console.error('No shapes selected.');
+        return;
+      }
+
+      const originalShape = selectedShapes[0];
+      console.log('Original shape:', originalShape);
+      const originalPosition = { x: originalShape.x, y: originalShape.y };
+      const originalWidth = originalShape.props.width || 200;
+      const originalHeight = originalShape.props.height || 50;
+      const offsetX = 20;
+      const offsetY = 50;
+
+      const createShapesPromises = items.map((item, index) => {
+        const position = {
+          x: originalPosition.x + originalWidth + offsetX,
+          y: originalPosition.y + index * (originalHeight + offsetY),
+        };
+
+        console.log('Creating shape at position:', position, 'with text:', item);
+        const newTextShapeId = createShapeId();
+
+        return editor.createShapes([
+          {
+            id: newTextShapeId,
+            type: 'text',
+            x: position.x,
+            y: position.y,
+            props: { text: item },
+          },
+        ]);
+      });
+
+      await Promise.all(createShapesPromises);
+      console.log('All shapes created');
+    } else {
+      console.error('No items generated.');
+    }
+  } catch (error) {
+    console.error('Error generating items:', error);
+  }
+};
+
+// Component to get selected texts and handle events
 const GetSelectedTexts = track(() => {
   const editor = useEditor();
+  const [selectedText, setSelectedText] = useState('');
+  const [selectedTexts, setSelectedTexts] = useState([]);
 
-  // Function to get selected texts
-  const selectedTexts = editor.getSelectedShapes()
-    .filter(shape => shape.type === 'text' && shape.props && shape.props.text)
-    .map(shape => shape.props.text);
+  // Function to update selected text
+  const updateSelectedText = () => {
+    const selectedTextsArray = editor.getSelectedShapes()
+      .filter(shape => shape.type === 'text' && shape.props && shape.props.text)
+      .map(shape => shape.props.text);
 
-  console.log('Filtered Selected Texts:', selectedTexts);
+    console.log('Updating selected text from selectedTexts:', selectedTextsArray);
+    if (selectedTextsArray.length > 0) {
+      setSelectedText(selectedTextsArray.join(' '));
+      setSelectedTexts(selectedTextsArray);  // Update the selectedTexts state
+      console.log('Selected text set to:', selectedTextsArray.join(' '));
+    } else {
+      setSelectedText('');
+      setSelectedTexts([]);  // Clear the selectedTexts state
+      console.log('Selected text cleared');
+    }
+  };
+
+  // Handle click event
+  const handleClick = (event) => {
+    console.log('Handling click at point:', { x: event.clientX, y: event.clientY });
+    const shape = editor.getShapeAtPoint({ x: event.clientX, y: event.clientY });
+    if (shape && shape.type === 'text') {
+      editor.selectShape(shape.id);
+      updateSelectedText(); // Update the selected text immediately after selection
+      console.log('Shape selected and text set:', shape.props.text);
+    } else {
+      console.log('No text shape found at clicked point');
+    }
+  };
+
+  useEffect(() => {
+    // Ensure editor is ready before adding listeners
+    if (!editor) return;
+    updateSelectedText();
+  }, [editor]);
+
+  useEffect(() => {
+    if (!editor) return;
+
+    console.log('Adding event listeners');
+    editor.on('pointerdown', handleClick);
+    editor.on('selectionchange', updateSelectedText);
+
+    return () => {
+      console.log('Removing event listeners');
+      editor.off('pointerdown', handleClick);
+      editor.off('selectionchange', updateSelectedText);
+    };
+  }, [editor]);
 
   return (
     <div style={{ display: 'none' }}>
@@ -22,3 +128,4 @@ const GetSelectedTexts = track(() => {
 });
 
 export default GetSelectedTexts;
+export { handleGenerateItems };
